+++ src/main/java/jira6/fate/domain/board/entity/Board.java
@@ -11,6 +11,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import jira6.fate.domain.user.entity.User;
 import jira6.fate.global.entity.Timestamped;
@@ -39,8 +40,11 @@ public class Board extends Timestamped {
     @JoinColumn(name = "user_id")
     private User user;
 
-    @OneToMany(mappedBy = "board", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Set<Columns> columns = new HashSet<>();
+    @OneToMany(mappedBy = "board", cascade = CascadeType.ALL)
+    private List<Column> columns;
+
+    @OneToMany(mappedBy = "board", cascade = CascadeType.ALL)
+    private List<Team> teams;
 
     @Builder
     public Board(String title, String intro) {
