+++ src/main/java/jira6/fate/domain/board/entity/Team.java
@@ -1,5 +1,6 @@
-package jira6.fate.domain.team.entity;
+package jira6.fate.domain.board.entity;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -7,30 +8,28 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import jira6.fate.domain.user.entity.User;
+import jira6.fate.global.entity.Timestamped;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "team_table")
-@Getter
-@NoArgsConstructor
-public class TeamUser {
+@Table(name = "table_team")
+public class Team extends Timestamped {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
 
     @ManyToOne
-    @JoinColumn(name = "team_id")
-    private Team team;
-
-    public TeamUser(User user, Team team) {
-        this.user = user;
-        this.team = team;
-    }
+    @JoinColumn(name = "board_id")
+    private Board board;
 }
