--- src/main/java/jira6/fate/domain/team/repository/TeamRepository.java
@@ -1,8 +1,0 @@
-package jira6.fate.domain.team.repository;
-
-import jira6.fate.domain.team.entity.Team;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<Team, Long> {
-
-}
