--- src/main/java/jira6/fate/domain/team/repository/UserTeamRepository.java
@@ -1,8 +1,0 @@
-package jira6.fate.domain.team.repository;
-
-import jira6.fate.domain.team.entity.TeamUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserTeamRepository extends JpaRepository<TeamUser, Long> {
-
-}
