--- src/main/java/jira6/fate/domain/team/service/TeamService.java
@@ -1,16 +1,0 @@
-package jira6.fate.domain.team.service;
-
-import jira6.fate.domain.team.entity.Team;
-import jira6.fate.domain.team.repository.TeamRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class TeamService {
-
-    private final TeamRepository teamRepository;
-
-    private final Te
-
-}
