+++ src/main/java/jira6/fate/global/config/SecurityConfig.java
@@ -2,9 +2,9 @@ package jira6.fate.global.config;
 
 import jira6.fate.domain.user.repository.UserRepository;
 import jira6.fate.global.jwt.JwtProvider;
+import jira6.fate.global.security.UserDetailsServiceImpl;
 import jira6.fate.global.security.filter.JwtAuthenticationFilter;
 import jira6.fate.global.security.filter.JwtAuthorizationFilter;
-import jira6.fate.global.security.UserDetailsServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
@@ -29,7 +29,8 @@ public class SecurityConfig {
     private final UserDetailsServiceImpl userDetailsService;
 
     @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration)
+        throws Exception {
         return configuration.getAuthenticationManager();
     }
 
@@ -51,13 +52,20 @@ public class SecurityConfig {
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
 
         http.csrf(AbstractHttpConfigurer::disable);
-        http.sessionManagement( (sessionManagement) -> sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authorizeHttpRequests( (authorizeHttpRequests) -> authorizeHttpRequests
-                        .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll()
-                        .requestMatchers("/api/users/signup").permitAll()
-                        .anyRequest().authenticated())
-                .addFilterBefore(jwtAuthorizationFilter(), JwtAuthenticationFilter.class)
-                .addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+        http.sessionManagement((sessionManagement) -> sessionManagement.sessionCreationPolicy(
+                SessionCreationPolicy.STATELESS))
+            .authorizeHttpRequests((authorizeHttpRequests) -> authorizeHttpRequests
+                .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll()
+                .requestMatchers("/api/users/signup").permitAll()
+                .requestMatchers(
+                    "/v2/api-docs", "/v3/api-docs", "/v3/api-docs/**", "/swagger-resources",
+                    "/swagger-resources/**", "/configuration/ui", "/configuration/security",
+                    "/swagger-ui/**",
+                    "/webjars/**", "/swagger-ui.html", "/api/auth/**", "/login", "/login.html"
+                ).permitAll()
+                .anyRequest().authenticated())
+            .addFilterBefore(jwtAuthorizationFilter(), JwtAuthenticationFilter.class)
+            .addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
     }
