--- src/main/java/jira6/fate/global/config/SwaggerConfig.java
@@ -1,19 +1,0 @@
-package jira6.fate.global.config;
-
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import io.swagger.v3.oas.models.*;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-            .info(new Info()
-                .title("fate API")
-                .description("팀 프로젝트 일정관리")
-                .version("1.0"));
-    }
-}
